+++ src/main/java/org/is/bandmanager/dto/request/BestBandAwardRequest.java
@@ -11,7 +11,7 @@ import org.is.bandmanager.model.MusicGenre;
 public class BestBandAwardRequest {
 
     @NotNull(message = "BestBandAward.MusicBandId не может быть пустым")
-    private Integer musicBandId;
+    private Long musicBandId;
 
     @NotNull(message = "BestBandAward.MusicGenre не может быть пустым")
     private MusicGenre genre;
